// app/backend/src/Interfaces/Leaderboards/Leaderboards.ts
export interface ILeaderboardEntry {
  name: string;
  totalPoints: number;
  totalGames: number;
  totalVictories: number;
  totalDraws: number;
  totalLosses: number;
  goalsFavor: number;
  goalsOwn: number;
  goalsBalance: number,
  efficiency: number,
}

export interface ILeaderboardParams {
  id: number;
  homeTeamId: number;
  homeTeamGoals: number;
  awayTeamId: number;
  awayTeamGoals: number;
  inProgress: boolean;
  homeTeam?: {
    id: number;
    teamName: string;
  };
}